hack out cmd-db remapping so it probes
the cmd-db driver is required to probe the rpmh driver, but the rpmh
driver doesn't use cmd-db directly. Instead each sub-device driver uses
cmd-db. We don't really need to use it, but we want to probe the rpmh
driver so that the clk-stub driver can probe and the mmc device can
probe without having to stub out the rpmh clk it uses.